backend/requirements.txt
langgraph
langchain-google-genai
langchain-core
python-dotenv
pydantic
pyyaml
numpy
langchain-huggingface
sentence-transformers